+++ lib/index.ts
@@ -1,3 +1,3 @@
 export * from './client/redis-stream.client';
-export * from './client/redis-stream.client.module';
+export * from './module/redis-stream.client.module';
 export * from './server/redis-stream.server';
