+++ lib/interface/redis-stream.interface.ts
