+++ lib/interface/redis-stream.serializtaion.interface.ts
